demo1/requirements.txt
Django>=1.10,<1.11
pytest-splinter==1.8.5
pytest-django==3.1.2
